backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
bleach==6.2.0
boto3==1.37.28
botocore==1.37.28
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-api-logger==1.1.16
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
